Significant Patch Panel changes ****READ****
Because we are changing how the robotic arm is powered, (now it is only
12V), i removed the robotic arm bus, It will just use a normal 12V high
power bus. The 12V high power bus is cleaned up a little bit.

I shrank down the planes to the areas they *need* to cover. Because of
the issues with the Powerboard, I’m less of a fan now of ground planes
everywhere.
The robotic arm bus is replaced with a 29V low-power (~3A) pass thru,
in case we need that sort of thing.
I also changed the way PoE injection is handled. This new
implementation will prevent gigabit pass thru, but gigabit pass thru
will be prevented by our implementation of PoE. Cutting the power line
to the switch helps limit problems.
I added dual-layer planes for all of the motor busses.