Create MRDT0003C1.sch
Schematic for a free running version of the active balancer designed for the slave BPS. The active balancer is purely analog and does not require any dedicated control/programing. It naturally balances the battery voltages when connected to two packs. Max voltage is 5V. Max battery mismatch should be limited to 50% to prevent overheating of the balancer. R2 can be replaced with a PTC resistor to clamp the balancer off if it overheats.