added file auxiliary_328_swser to allow comm with GPS via SW Serial
Okay, so, with the previous copy of the GPS code (auxiliary_328), I had
difficulty showing that I was getting the correct values via the
Software Serial UARTS, however, Hardware Serial UARTS worked fine. The
reason this is important is because our custom 32U4 boards do not
currently have working Hardware Serial UARTS (we don't know why). It
seems as though the problem was caused by a combination of (a) a lack of
delays following Serial (SW and HW) begin functions and (b) the clock
edits done in the previous version of Setup.h. Tl;dr auxiliary_328
contains the code for communicating with GPS if we get Hardware Serial
UARTS working on the custom 32U4 boards, auxiliary_328_swser contains
the code for communicating with GPS if we must resort to using Software
Serial UARTS. Now, because it is Finals Week, and I hate this code, I'm
going home and going to bed. Goodnight!