working on sonar - serial not working - pre-pwm switch
I have written code that should be receiving serial data from the sonar
sensor however the rx is littered with framing errors and garbage data
(I logic analyzed it). I can also not get the sonar pulse pin to pull
low - it is pulled high on the sonar breakout and the atmega32u4 cannot
pull it low when powered off USB or OKI3V3 SMPS. I am going to switch to
code that reads sonar data via the pwm functionality.