The motor controller driver for powerboard is ready for rev 1
This is the complete powerboard design based off of what was decided at the Gimbal meetings. The actual schematic for the board is simply 1A.sch and the board design is 1A.brd I did change some pin assignment around from the orginal. Mainly moving 12V_ACT to the other pin board. Other than that the .sch file is unalterred. The board is grounded on top and bottom. Large polygons take are on the buses and mosfets to transmit more power and not fry wires.  There are no holes yet in the corners and they are to be added at a later date.