Final simplification of TI's F28069M microcontroller
Decided to remove all muxes and logic gates and make direct connection to GPIO 28 and 29 pins. No need to use their other functionality, serial TX and RX. We believe those serial lines are not involved when identifing a motor or flashing the microcontroller. We also decided on just one switch on TRST signal. The state of that line should decide whether code is automatically run on power up or if code is being flashed.