Rearanged Board and rounded edges
Put Anderson Connectors in correct places as well as rewired the whole board. I then rounded the corners of the board and put wholes in the board so we could properly mount it. Eagle would not allow .144 inch holes to be drilled, but .143985 inch holes. Should be precise enough.. The headers are now spaced correctly apart and there should be enough space for the control card to sit on top of it now. Polygons were also redrawn and the board is much smaller now.  It passed the design rules check as well, so it should be almost ready. Whenever Control card is finished, pin assignments may need to be changed, other than that, the board could technically go into production.