Added Silkscreen and Bill of Materials
Bill of materials is now added and can be changed as time goes on. All objects on board have been smashed and names are fairly visible. Some areas around phase C are hard to read due to spacing. Silkscreen is mostly finished. Added Logo to schematic so that it could be placed on the board. Almost all testpoints are together for easy probing. They are all on the Left side of the board along the pin heads and 3 are seperated off from the group due to spacing issues.