Put Max232 and related components on board
Tiva Chip is still in two components and so its routing remains to be done. The Tiva Headers are now underneath the board and all connectors are on one side of the board. The routing has been done for all components that are not connected to the Tiva chip have been routed. Will most likely need to be redone when chip is one full piece. Components are placed and ready to go.