Finished Routing Board
Board is finished with routing, however it needs silkscreen to be done. Silk screen will be completed at a later date. There is one unbroken ground plain. Top plain is fairly broken as a ground plane, still considerred ground with via patching. Bottom is about the same as Top, but less broken. The Upper Middle is mostly unbroken and and the Lower Middle plain is unbroken.